+++ packages/jestscreenshot/rollup.config.js
@@ -1,0 +1,7 @@
+import configure from '../../rollup.config';
+import builtins from 'rollup-plugin-node-builtins';
+
+export default configure({
+  input: 'src/index.ts',
+  externalConfig: builtins,
+});